separate vehicle domain
