+++ src/main/kotlin/com/micartel/telemetry/domain/model/CannotRefuelElectricVehicleException.kt
@@ -1,4 +1,7 @@
 package com.micartel.telemetry.domain.model
 
-data class CannotRefuelElectricVehicleException(val chassis: ChassisNumber, val category:Category)
+import com.micartel.telemetry.domain.model.vehicle.Category
+import com.micartel.telemetry.domain.model.vehicle.ChassisNumber
+
+data class CannotRefuelElectricVehicleException(val chassis: ChassisNumber, val category: Category)
     : RuntimeException("Cannot refuel electric vehicle $chassis category $category")
