--- src/main/kotlin/com/micartel/telemetry/domain/model/UniqueVehicleValidator.kt
@@ -1,5 +1,0 @@
-package com.micartel.telemetry.domain.model
-
-interface UniqueVehicleValidator {
-    fun validate(vehicle:Vehicle)
-}
