+++ src/main/kotlin/com/micartel/telemetry/domain/model/vehicle/LicensePlate.kt
@@ -1,4 +1,4 @@
-package com.micartel.telemetry.domain.model
+package com.micartel.telemetry.domain.model.vehicle
 
 val licensePlatePattern = "[A-Z0-9]+".toRegex()
