+++ src/main/kotlin/com/micartel/telemetry/domain/model/vehicle/UniqueVehicleValidator.kt
@@ -1,0 +1,5 @@
+package com.micartel.telemetry.domain.model.vehicle
+
+interface UniqueVehicleValidator {
+    fun validate(vehicle: Vehicle)
+}
