+++ src/main/kotlin/com/micartel/telemetry/domain/model/vehicle/VehicleAlreadyExistsException.kt
@@ -1,4 +1,4 @@
-package com.micartel.telemetry.domain.model
+package com.micartel.telemetry.domain.model.vehicle
 
 data class VehicleAlreadyExistsException(val chassis: ChassisNumber)
     : RuntimeException("Vehicle $chassis already exists")
