+++ src/main/kotlin/com/micartel/telemetry/port/adapter/persistence/inmemory/InMemoryUniqueVehicleValidator.kt
@@ -1,6 +1,8 @@
 package com.micartel.telemetry.port.adapter.persistence.inmemory
 
-import com.micartel.telemetry.domain.model.*
+import com.micartel.telemetry.domain.model.vehicle.UniqueVehicleValidator
+import com.micartel.telemetry.domain.model.vehicle.Vehicle
+import com.micartel.telemetry.domain.model.vehicle.VehicleAlreadyExistsException
 import org.springframework.stereotype.Component
 
 @Component
